requirements.txt
pandas
sqlalchemy
logging
pytest
psycopg2-binary
setuptools